+++ src/extension.ts
@@ -1,10 +1,12 @@
 /* eslint-disable import/prefer-default-export */
+import vscode from 'vscode';
 
-import * as vscode from 'vscode';
 import commands from './commands';
 
 export function activate(context: vscode.ExtensionContext) {
-    commands.forEach(command =>
-        context.subscriptions.push(vscode.commands.registerCommand(command.identifier, command.handler))
+    commands.forEach((command) =>
+        context.subscriptions.push(
+            vscode.commands.registerCommand(command.identifier, command.handler),
+        ),
     );
 }