+++ src/util.ts
@@ -1,6 +1,5 @@
-import * as vscode from 'vscode';
+import vscode from 'vscode';
 import axios from 'axios';
-import { get } from 'dot-prop';
 import packageJson, { FullMetadata } from 'package-json';
 
 /**
@@ -20,10 +19,10 @@ export async function fetchNpmPackageRepository(moduleName: string): Promise<str
             return null;
         }
 
-        const homepage = get<string>(metadata, 'homepage');
-        if (homepage && homepage.startsWith('https://github.com/')) return homepage;
+        const homepage = metadata?.homepage;
+        if (homepage?.startsWith('https://github.com/')) return homepage;
 
-        const repositoryURL: string | undefined = get<string>(metadata, 'repository.url');
+        const repositoryURL: string | undefined = metadata?.repository?.url;
         if (repositoryURL) {
             if (repositoryURL.startsWith('https://github.com/')) {
                 return repositoryURL;
@@ -44,7 +43,7 @@ export async function fetchNpmPackageRepository(moduleName: string): Promise<str
         return null;
     }
 
-    const repositoryURL: string | undefined = get(resp, 'data.collected.metadata.links.repository');
+    const repositoryURL: string | undefined = resp?.data?.collected?.metadata?.links?.repository;
     if (repositoryURL) {
         return repositoryURL;
     }
@@ -59,26 +58,30 @@ export function extractModuleNames(textContent: string): string[] {
     const exportRegexp = new RegExp(`export\\s+.*?from\\s+${moduleNameRegexpSource}`);
     const importStatementRegexp = new RegExp(
         `${requireRegexp.source}|${importRegexp.source}|${exportRegexp.source}`,
-        'g'
+        'g',
     );
     const importStatements = textContent.match(importStatementRegexp);
 
     if (importStatements) {
         return importStatements
-            .map(importStatement => {
-                const matchedRegexp = [requireRegexp, importRegexp, exportRegexp].find(regexp =>
-                    regexp.test(importStatement)
+            .map((importStatement) => {
+                const matchedRegexp = [requireRegexp, importRegexp, exportRegexp].find((regexp) =>
+                    regexp.test(importStatement),
                 )!;
                 const moduleName = importStatement.match(matchedRegexp)![2];
+                const firstSlashIndex = moduleName.indexOf('/');
+                if (~firstSlashIndex) {
+                    if (moduleName.startsWith('@')) {
+                        const secondSlashIndex = moduleName.indexOf('/', firstSlashIndex + 1);
+                        if (~secondSlashIndex) return moduleName.slice(0, secondSlashIndex);
+                    }
 
-                if (!moduleName.startsWith('@')) {
-                    const slashIndex = moduleName.indexOf('/');
-                    if (~slashIndex) return moduleName.slice(0, slashIndex);
+                    return moduleName.slice(0, firstSlashIndex);
                 }
 
                 return moduleName;
             })
-            .filter(moduleName => !moduleName.startsWith('@types/'));
+            .filter((moduleName) => !moduleName.startsWith('@types/'));
     }
 
     return [];
@@ -113,5 +116,5 @@ export function getPackageNamesFromPackageJSON(jsonTextContent: string): string[
 
     if (packageJSON.devDependencies) packageNames.push(...Object.keys(packageJSON.devDependencies));
 
-    return packageNames.filter(packageName => !packageName.startsWith('@types/'));
+    return packageNames.filter((packageName) => !packageName.startsWith('@types/'));
 }